requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiohttp-swagger==1.0.16
aiosignal==1.3.1
async-lru==2.0.2
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
cryptography==40.0.2
dill==0.3.6
distro==1.8.0
fire==0.5.0
frozenlist==1.3.3
genie==23.2
genie.libs.clean==23.2
genie.libs.conf==23.2
genie.libs.filetransferutils==23.2
genie.libs.health==23.2
genie.libs.ops==23.2
genie.libs.parser==23.2
genie.libs.sdk==23.2
gitdb==4.0.10
GitPython==3.1.31
grpcio==1.54.0
idna==3.4
Jinja2==3.1.2
jsonpickle==3.0.1
junit-xml==1.9
lxml==4.9.2
MarkupSafe==2.1.2
multidict==6.0.4
ncclient==0.6.13
netaddr==0.8.0
packaging==23.1
paramiko==3.1.0
pathspec==0.11.1
prettytable==3.7.0
protobuf==3.20.3
psutil==5.9.5
pyats==23.2
pyats-genie-command-parse==1.3.5
pyats.aereport==23.2
pyats.aetest==23.2
pyats.async==23.2
pyats.connections==23.2
pyats.contrib==23.4
pyats.datastructures==23.2
pyats.easypy==23.2
pyats.kleenex==23.2
pyats.log==23.2
pyats.reporter==23.2
pyats.results==23.2
pyats.tcl==23.2
pyats.topology==23.2
pyats.utils==23.2
pycparser==2.21
pyftpdlib==1.5.7
PyNaCl==1.5.0
python-engineio==3.14.2
python-socketio==4.6.1
PyYAML==6.0
requests==2.28.2
requests-toolbelt==1.0.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
smmap==5.0.0
termcolor==2.3.0
tftpy==0.8.0
tqdm==4.65.0
typing-extensions==4.5.0
unicon==23.2
unicon.plugins==23.2
urllib3==1.26.15
wcwidth==0.2.6
xlrd==1.2.0
XlsxWriter==3.1.0
xlwt==1.3.0
xmltodict==0.13.0
yamllint==1.31.0
yang.connector==23.3
yarl==1.9.1
